backend/requirements.txt
annotated-types==0.7.0
anyio==4.5.0
appnope==0.1.4
asttokens==2.4.1
backcall==0.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.8.0.post1
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
google-ai-generativelanguage==0.1.0
google-api-core==2.20.0
google-auth==2.35.0
google-generativeai==0.1.0rc1
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.62.3
grpcio-tools==1.66.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
hyperframe==6.0.1
idna==3.10
importlib_resources==6.4.5
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
packaging==24.1
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
portalocker==2.10.1
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.11.3
regex==2024.9.11
requests==2.32.3
rsa==4.9
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
sentence-transformers==3.1.1
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.0
torch==2.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
zipp==3.20.2
